pyproject.toml
[project]
name = "skyengine"
version = "0.1.0"
description = "天工的配置文件表"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.125.0",
    "gymnasium>=0.28.1",
    "matplotlib>=3.10.8",
    "opencv-python>=4.8.1.78",
    "ortools>=9.14.6206",
    "pettingzoo>=1.24.3",
    "pogema==1.3.1",
    "pygame>=2.6.1",
    "pygame-gui>=0.6.14",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "uvicorn>=0.40.0",
]

[tool.uv]
override-dependencies = ["pydantic>=1.10.13,<2.0.0"]

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
torchvision = { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" }
